*REALLY* solve the mystery-win32-failure cpantesters report
While fa19e5d6 was a nice idea, its test was flawed. As seen in the fresh fail
http://www.cpantesters.org/cpan/report/137dfaf3-7bbc-1014-aec2-3d202b825c07
the writability test passed with flying colors, but the lock didn't open right
after that. The only explanation at this point is that the FS underlying this
particular dir can not create filenames starting with a dot <facepalm />

Thus (as unlikely as it seems) the culprit has been identified and the names
changed and they smoked happily ever after...